Bugfix v1.21.2

+++ src/utils-exported.cpp
@@ -293,16 +293,16 @@ Rcpp::String collapse_cpp(const Rcpp::StringVector &x) {
 // [[Rcpp::export(rng = false)]]
 void print_pb(const R_xlen_t &out) {
   if (out >= 10 && out < 100) {
-    REprintf("\b\b\b\b %i%%", out);
+    REprintf("\b\b\b\b %i%%", (int) out);
     return;
   }
   if (out > 0 && out < 10) {
-    REprintf("\b\b\b\b  %i%%", out);
+    REprintf("\b\b\b\b  %i%%", (int) out);
     return;
   }
   switch (out) {
     case   0: REprintf("   0%%");            return;
-    case 100: REprintf("\b\b\b\b%i%%", out); return;
+    case 100: REprintf("\b\b\b\b%i%%", (int) out); return;
     case  -1: REprintf("\b\b\b\b100%%\n");   return;
   }
   Rcpp::stop("Input must be an integer in between -1 and 100");
